requirements.txt
PyYAML~=6.0.1
opencv-python~=4.7.0.72
ultralytics~=8.1.16